# backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/services": {
			"get": {
				"operationId": "GetAllServices",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"icon": {
												"type": "string"
											},
											"description": {
												"type": "string"
											},
											"title": {
												"type": "string"
											},
											"createdAt": {
												"type": "string",
												"format": "date-time"
											},
											"image": {
												"type": "string"
											},
											"id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"icon",
											"description",
											"title",
											"createdAt",
											"image",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateService",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"service": {
											"properties": {
												"icon": {
													"type": "string"
												},
												"description": {
													"type": "string"
												},
												"title": {
													"type": "string"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"image": {
													"type": "string"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"icon",
												"description",
												"title",
												"createdAt",
												"image",
												"id"
											],
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"service",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"image": {
										"type": "string"
									},
									"icon": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"title": {
										"type": "string"
									}
								},
								"required": [
									"icon",
									"description",
									"title"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/services/{id}": {
			"get": {
				"operationId": "GetServiceById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"icon": {
													"type": "string"
												},
												"description": {
													"type": "string"
												},
												"title": {
													"type": "string"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"image": {
													"type": "string"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"icon",
												"description",
												"title",
												"createdAt",
												"image",
												"id"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateService",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"service": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"service": {
													"properties": {
														"icon": {
															"type": "string"
														},
														"description": {
															"type": "string"
														},
														"title": {
															"type": "string"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"image": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"icon",
														"description",
														"title",
														"createdAt",
														"image",
														"id"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"service",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"image": {
										"type": "string"
									},
									"icon": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"title": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteService",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}